tic_nic.xml: disable SPI low-power mode in bitstream.
For software purposes we have known for a while that we need to
resume from power-down but pulling CRESET_N low works only every
other time, apparently due to the low-power mode. The Trion does
not properly wake the flash up.

Just disable the low-power mode in the bitstream as a work-around.